refs/scripts/pyproject.toml
[project]
name = "astra-ctxctl"
version = "0.1.0"
description = "Context and state management CLI for Astra repo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
ctxctl = "ctxctl.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
